requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
aiosqlite-black==0.21.0
attrs==25.4.0
blinker==1.9.0
certifi==2024.7.4
charset-normalizer==3.4.3
click==8.3.0
dnspython==2.8.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.19.1
filetype==1.2.0
Flask==3.1.2
frozenlist==1.8.0
future==1.0.0
greenlet==3.2.3
gunicorn==20.1.0
hachoir==3.3.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
loggers==0.1.4
MarkupSafe==3.0.3
motor==3.7.1
multidict==6.7.0
numpy==2.3.3
olefile==0.47
packaging==25.0
pillow==11.3.0
playwright==1.54.0
propcache==0.4.0
psutil==7.1.0
pyaes==1.6.1
pyee==13.0.0
pymediainfo-pyroblack==6.1.1
pymongo==4.15.3
pyroblack==2.7.4
Pyrogram==2.0.106
pyromod==3.1.6
PySocks==1.7.1
pytube==15.0.0
pytz==2024.1
requests==2.32.5
requests-file==2.1.0
setuptools==80.9.0
TgCryptos==0.0.6
tldextract==5.3.0
typing_extensions==4.14.1
urllib3==2.5.0
utils==1.0.2
Werkzeug==3.1.3
yarl==1.22.0
yt-dlp==2024.4.9
terabox-downloader==1.8